refactor(hub): move policy/ into kernel/ (Stage 3c)
Pure git mv + relative-import depth fixes; no logic change.
