+++ packages/hub/src/kernel/policy/noydb-facade.ts
@@ -16,13 +16,13 @@
  *
  * Internal subsystem — reached through `noydb.getPolicy(...)` etc.
  */
-import { ValidationError } from '../kernel/errors.js'
+import { ValidationError } from '../errors.js'
 import { PERSONAL_POLICY, mergePolicy } from './presets.js'
 import { loadVaultPolicy, saveVaultPolicy } from './storage.js'
 import type { VaultPolicy } from './types.js'
-import type { NoydbStore, ReAuthOperation, SessionPolicy } from '../kernel/types.js'
-import type { PolicyEnforcer } from '../with-party/session/session-policy.js'
-import type { SessionStrategy } from '../with-party/session/strategy.js'
+import type { NoydbStore, ReAuthOperation, SessionPolicy } from '../types.js'
+import type { PolicyEnforcer } from '../../with-party/session/session-policy.js'
+import type { SessionStrategy } from '../../with-party/session/strategy.js'
 
 /** Everything the moving policy/session methods touched on the Noydb instance's `this.*`. */
 export interface NoydbPolicyDeps {
