+++ packages/hub/src/kernel/policy/storage.ts
@@ -8,8 +8,8 @@
  *
  * @module
  */
-import type { NoydbStore, EncryptedEnvelope } from '../kernel/types.js'
-import { NOYDB_FORMAT_VERSION } from '../kernel/types.js'
+import type { NoydbStore, EncryptedEnvelope } from '../types.js'
+import { NOYDB_FORMAT_VERSION } from '../types.js'
 import type { VaultPolicy } from './types.js'
 
 /** Reserved collection name for vault-level metadata documents. */
